Make moe slightly more conventioanl wrt starting a wsgi app and using cli actions module
